Some refactoring and precision

+++ src/main/java/logged/DeleteAssociationServlet.java
@@ -1,6 +1,7 @@
 package logged;
 
 import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import dao.DAO;
 import utils.Useful;
 
@@ -9,6 +10,7 @@ import javax.servlet.http.*;
 import javax.servlet.annotation.*;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.lang.reflect.Type;
 import java.sql.SQLException;
 
 @WebServlet(name = "DeleteAssociationServlet", urlPatterns = "/DeleteAssociationServlet")
@@ -38,35 +40,47 @@ public class DeleteAssociationServlet extends HttpServlet {
         HttpSession session = request.getSession(false);
         PrintWriter out = response.getWriter();
 
-        String opCode = request.getParameter("opCode");
         Gson gson = new Gson();
         String Json;
 
         int idDocente = 0;
         String nomeCorso = "";
 
-        Useful message;
+        Useful message = new Useful();
 
         if (session != null) {
             String sessionId = request.getParameter("jSessionId");
             if (sessionId.equals(session.getId())) {
+                if (session.getAttribute("ruolo").equals("admin")) {
 
-                nomeCorso = request.getParameter("nomeCorso");
-                idDocente = Integer.parseInt(request.getParameter("idDocente"));
-                try {
-                    dao.deleteCorsoDocenteAssociation(nomeCorso, idDocente);
-                    message = new Useful("Successfully deleted association", 1, null);
-                    Json = gson.toJson(message);
-                    out.write(Json);
-                    out.flush();
-                } catch (SQLException e) {
-                    System.out.println(e.getMessage());
-                    message = new Useful("Unsuccessfully deleted association", -1, null);
-                    Json = gson.toJson(message);
-                    out.write(Json);
-                    out.flush();
+                    nomeCorso = request.getParameter("nomeCorso");
+                    idDocente = Integer.parseInt(request.getParameter("idDocente"));
+                    try {
+                        dao.deleteCorsoDocenteAssociation(nomeCorso, idDocente);
+                        message = new Useful("Successfully deleted association", 1, null);
+
+                    } catch (SQLException e) {
+                        System.out.println(e.getMessage());
+                        message = new Useful("Unsuccessfully deleted association", -1, null);
+                        Type type = new TypeToken<Useful>() {
+                        }.getType();
+                        Json = gson.toJson(message, type);
+                        out.write(Json);
+                        out.flush();
+                    }
                 }
             }
+            else {
+                message = new Useful("Sorry but your id doesn't correspond", -1, null);
+            }
         }
+        else {
+            message = new Useful("Sorry but you don't appear to be logged in", -1, null);
+        }
+        Type type = new TypeToken<Useful>() {
+        }.getType();
+        Json = gson.toJson(message, type);
+        out.write(Json);
+        out.flush();
     }
 }
